+++ src/font_Ctx.ts
@@ -1,2 +1,0 @@
-import type { font_ctx_I } from './font_ctx_I.generated.js'
-export interface font_Ctx extends font_ctx_I {}