+++ src/index.ts
@@ -1,5 +1,3 @@
-export * from './font_Ctx.js'
-export * from './font_ctx_I.generated.js'
 export * from './clone_ctx_T.js'
 export * from './fontSize_downscale_fit.js'
-export * from './rem_px_b.js'
+export * from './rem_px$_b.js'